pyproject.toml
[project]
name = "chronocare"
version = "0.1.0"
description = "老年父母健康管理平台"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "fastapi>=0.136.1",
    "greenlet>=3.5.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.14.0",
    "python-multipart>=0.0.27",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"src/chronocare/services/notification.py" = ["E501"]

[tool.mypy]
python_version = "3.14"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
